refactor: ubdate file structure
