+++ server/controllers/createMatch.ts
@@ -1,13 +1,13 @@
 import { Request, Response, NextFunction, RequestHandler } from 'express';
-import { createMatchService } from '../../services';
-import CustomRequest from '../../interfaces';
+import { createMatchService } from '../services';
+import CustomRequest from '../interfaces';
 
 const createMatch: RequestHandler = async (
   req: CustomRequest,
   res: Response,
   next: NextFunction,
 ): Promise<void> => {
-  req.userData = { owner_id: 1 }; // this id will coming from checkAuth middleware
+  req.userData = { owner_id: 1 };
   const data = await createMatchService(req);
   res.json({
     status: 202,
